Set up recipe card partials
