+++ src/views/dashboard.handlebars
@@ -1,4 +1,22 @@
 {{> privateNav}}
 <div class="flex justify-center my-4">
-  <div class="font-bold text-xl mb-2">Welcome {{email}}</div>
+  <form class="w-full max-w-sm" action="/recipes" method="POST">
+    <div class="flex items-center border-b border-b-2 border-teal-500 py-2">
+      <input
+        class="appearance-none bg-transparent border-none w-full text-gray-700 mr-3 py-1 px-2 leading-tight focus:outline-none"
+        name="searchKeyword" type="text" placeholder="Jane Doe" aria-label="Full name">
+      <button
+        class="flex-shrink-0 bg-teal-500 hover:bg-teal-700 border-teal-500 hover:border-teal-700 text-sm border-4 text-white py-1 px-2 rounded"
+        type="submit">
+        Search
+      </button>
+    </div>
+  </form>
+  <div>
+    {{#if recipes}}
+    {{#each recipes}}
+    {{> recipeCard}}
+    {{/each}}
+    {{/if}}
+  </div>
 </div>
