+++ src/views/partials/recipeCard.handlebars
@@ -1,0 +1,19 @@
+<div class="max-w-sm w-full lg:max-w-full lg:flex">
+    <div class="h-48 lg:h-auto lg:w-48 flex-none bg-cover rounded-t lg:rounded-t-none lg:rounded-l text-center overflow-hidden"
+        style="background-image: url('{{imageUrl}}')" title={{label}}>
+    </div>
+    <div
+        class="border-r border-b border-l border-gray-400 lg:border-l-0 lg:border-t lg:border-gray-400 bg-white rounded-b lg:rounded-b-none lg:rounded-r p-4 flex flex-col justify-between leading-normal">
+        <div class="mb-8">
+            <p class="text-sm text-gray-600 flex items-center">
+                Calories: {{calories}}
+            </p>
+            <div class="text-gray-900 font-bold text-xl mb-2">{{label}}</div>
+        </div>
+        <div class="flex items-center">
+            <div class="text-sm">
+                <p class="text-gray-900 leading-none">{{source}}</p>
+            </div>
+        </div>
+    </div>
+</div>
